AMP zip file
What is AMP?
AMP is an automated message player which can detect the face of the user and can play the message which is in list for that particular user only.
only the admin has the ability to add more users to  the system and record the message for them and save it in the list of audio messages saved.

Why Automatic Message Player?
We are in a country where an specially able person has to rely on other individuals for their basic need. Through this application we are intending to help blind and physically disabled person to at least be able to receive their messages without someones help.

AMP specifically works on c# and uses visual basic. It makes use of facial recognition techniques and algorithams along with hasscascade face detection methods.